Day 4: part 2 original solution
This felt better using more iter functions from rust, but splitting
still feels like I have some improvement. A way to ignore the extra
"  " or "    " chars as whitespace for a split would feel more comfortable.
Instead, i'm filtering checking that we aren't an empty item instead.
I'm not sure if this is the right/best way to solve that part of the parsing problem.